insurance-app/services/application_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg
pydantic==2.7.1
PyJWT==2.8.0
python-dotenv==1.0.1
motor==3.7.1
python-jose
python-jose[cryptography]==3.3.0
cryptography==42.0.5
nats-py